requirements.txt
# NyayMitra — Backend Dependencies
# Install: pip install -r requirements.txt

# ── Web Framework ──
fastapi==0.135.3
uvicorn==0.44.0

# ── Vector Database ──
chromadb==1.5.7

# ── Embeddings ──
sentence-transformers==5.4.0

# ── LLM Clients (Sarvam-M + Groq fallback) ──
openai==2.31.0

# ── HTTP Client (Sarvam Translate API) ──
httpx==0.28.1

# ── Utilities ──
python-dotenv==1.2.2
langdetect==1.0.9
pydantic==2.12.5